Regex soft-retweet detection
I added a regex to detect 'RT @' anywhere in the content of a tweet, in
case someone hides a RT in the middle of a tweet! As a bonus, it also
detects a bunch of different types of quotes as well as other things
people might use, like 'via,' 'by,' and 'from'!