Updated pictweet to use upload()
Twitter's website states that update_with_media is deprecated, so I
replaced it when their new API thingie, upload!

I upgraded it a little bit so that you can actually put multiple images
in a single tweet. We've been able to do that manually for a while now!

Hash parameters work a little differently now, since I made it so that
you can pass in two option hashes! The first one will be treated as
options for the posting your tweet, while the second one will be used as
options for your upload. I don't know if options even exist for that,
but I implemented them anyways just in case!